requirements.txt
Flask
gunicorn
numpy
grpcio
grpcio-tools
accelerate
diffusers
protobuf
torch
torchvision
transformers
pillow
boto3
botocore
